Continue organizing tests
